Address review comments from Alex
